Consolidated WeatherFrame into Weather
